add view student file and modified student file controller

+++ controllers/student/studentFileController.js
@@ -1,4 +1,4 @@
-const { Student_file } = require('../../models/index.js')
+const { Student_file, View_student_file } = require('../../models/index.js')
 const { resJSON, errorJSON } = require('../../repository/resJSON.js.js')
 const { deleteData, updateData, createData, getData } = require('../../repository/crudAction.js')
 const path = require('path')
@@ -7,18 +7,19 @@ const message = ' student file successfully'
 
 const getAllStudentFile = async (req, res) => {
   try {
-    const { id_study_year, id_class_name } = req.body
+    const { id_study_year, id_class_name, semester } = req.query
 
     if (id_study_year && id_class_name) {
-      const dataStudentFile = await Student_file.findAll({
+      const dataStudentFile = await View_student_file.findAll({
         where: {
           id_study_year: id_study_year,
-          id_class_name: id_class_name
+          id_class_name: id_class_name,
+          semester: semester
         }
       })
       resJSON(res, dataStudentFile, 'get'+message)
     } else {
-      const dataStudentFile = await getData(Student_file)
+      const dataStudentFile = await getData(View_student_file)
       resJSON(res, dataStudentFile, 'get'+message)
     }
 
@@ -31,21 +32,35 @@ const createStudentFile = async (req, res) => {
   try {
     const { id_student, id_study_year, id_class_name, file_name, semester, category } = req.body
 
-    if (!id_student || !id_student || !id_class_name || !file_name || !category || !req.file) {
-      return errorJSON(res, 'request has not been fulfilled', 400)
+    if (!id_student || !id_study_year || !category) {
+      return errorJSON(res, 'request has not been fulfilled', 406)
     } else {
-      const file_upload = req.file.path
 
-      await createData(Student_file, {
-        id_student: id_student,
-        id_study_year: id_study_year,
-        id_class_name: id_class_name,
-        file: file_upload,
-        file_name: file_name,
-        semester: semester,
-        category: category
-      })
-      resJSON(res, '', 'create'+message)
+      if (req.file) {
+        const file_upload = req.file.path
+
+        await createData(Student_file, {
+          id_student: id_student,
+          id_study_year: id_study_year,
+          id_class_name: id_class_name,
+          file: file_upload,
+          file_name: file_name,
+          semester: semester,
+          category: category
+        })
+        resJSON(res, '', 'create'+message)
+      } else {
+        await createData(Student_file, {
+          id_student: id_student,
+          id_study_year: id_study_year,
+          id_class_name: id_class_name,
+          file_name: file_name,
+          semester: semester,
+          category: category
+        })
+        resJSON(res, '', 'create'+message)
+      }
+      
     }
   } catch (error) {
     errorJSON(res)
@@ -54,7 +69,7 @@ const createStudentFile = async (req, res) => {
 
 const removeFile = (filePath) => {
   // get location image
-  filePath = path.join(__dirname, '..', filePath)
+  filePath = path.join(__dirname, '../..', filePath)
 
   // remove file by path
   fs.unlink(filePath, err => console.log(err, '<-- error remove file'))
@@ -66,8 +81,7 @@ const deleteStudentFile = async (req, res) => {
 
     const data = await deleteData(Student_file, id)
     if (!data) return errorJSON(res, 'data is not found', 404);
-    
-    removeFile(data.file)
+    if (data.file) removeFile(data.file)
 
     resJSON(res, '', 'delete'+message)
   } catch (error) {
@@ -77,10 +91,10 @@ const deleteStudentFile = async (req, res) => {
 
 const updateStudentFile = async (req, res) => {
   try {
-    const { id, id_student, file_name, classes, semester } = req.body
+    const { id, id_student, id_study_year, id_class_name, file_name, semester, category } = req.body
 
-    if (!id_student || !file_name || !classes || !semester) {
-      return errorJSON(res, 'request has not been fulfilled', 400)
+    if (!id_student || !id_study_year || !category) {
+      return errorJSON(res, 'request has not been fulfilled', 406)
     } else {
 
       const dataStudentFile = await Student_file.findOne({
@@ -89,7 +103,8 @@ const updateStudentFile = async (req, res) => {
   
       if (req.file) {
         const file_upload = req.file.path
-        if (dataStudentFile.file === '') {
+
+        if (!dataStudentFile.file) {
           dataStudentFile.file = file_upload
         } else {
           removeFile(dataStudentFile.file)
@@ -97,8 +112,7 @@ const updateStudentFile = async (req, res) => {
         }
       }
 
-      dataStudentFile.class = classes
-      await updateData(Student_file, id, { id_student, file_name, semester })
+      await updateData(Student_file, id, { id_student, id_study_year, id_class_name, file_name, semester, category })
 
       dataStudentFile.save()
